@@ -1,20 +1,21 @@
 name: signalr_core
 description: ASP.NET Core SignalR Client written for Dart that is platform-independent, and can be used on both the command-line and the browser.
-version: 1.0.8
+version: 1.1.0
+publish_to: none
 homepage: https://github.com/jamiewest/signalr_core
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.1
-  logging: ^0.11.4
-  meta: ^1.1.8
-  tuple: ^1.0.3
-  web_socket_channel: ^1.1.0
-  sse_client: ^0.0.3
-  equatable: ^1.2.3
+  http: ^0.13.0
+  logging: ^1.0.0
+  meta: ^1.3.0
+  tuple: ^2.0.0-nullsafety.0
+  web_socket_channel: ^2.0.0
+  sse_client: ^0.1.0
+  equatable: ^2.0.0
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.2.0
+  pedantic: ^1.11.0
+  test: ^1.16.5
