@@ -4,23 +4,23 @@ import 'package:http/http.dart';
 import 'package:signalr_core/signalr_core.dart';
 
 class LongPollingTransport implements Transport {
-  final BaseClient _client;
-  final AccessTokenFactory _accessTokenFactory;
-  final Logging _log;
-  final bool _logMessageContent;
-  final bool _withCredentials;
+  final BaseClient? _client;
+  final AccessTokenFactory? _accessTokenFactory;
+  final Logging? _log;
+  final bool? _logMessageContent;
+  final bool? _withCredentials;
 
-  String _url;
-  bool _running;
-  Future<void> _receiving;
-  Exception _closeError;
+  String? _url;
+  late bool _running;
+  Future<void>? _receiving;
+  Exception? _closeError;
 
   LongPollingTransport({
-    BaseClient client,
-    AccessTokenFactory accessTokenFactory,
-    Logging log,
-    bool logMessageContent,
-    bool withCredentials,
+    BaseClient? client,
+    AccessTokenFactory? accessTokenFactory,
+    Logging? log,
+    bool? logMessageContent,
+    bool? withCredentials,
   })  : _client = client,
         _accessTokenFactory = accessTokenFactory,
         _log = log,
@@ -32,16 +32,16 @@ class LongPollingTransport implements Transport {
   }
 
   @override
-  var onclose;
+  OnClose? onclose;
 
   @override
-  var onreceive;
+  OnReceive? onreceive;
 
   @override
-  Future<void> connect(String url, TransferFormat transferFormat) async {
+  Future<void> connect(String? url, TransferFormat? transferFormat) async {
     _url = url;
 
-    _log(LogLevel.trace, '(LongPolling transport) Connecting.');
+    _log?.call(LogLevel.trace, '(LongPolling transport) Connecting.');
 
     final headers = <String, String>{};
     final userAgentHeader = getUserAgentHeader();
@@ -55,10 +55,10 @@ class LongPollingTransport implements Transport {
     // Make initial long polling request
     // Server uses first long polling request to finish initializing connection and it returns without data
     final pollUrl = '$url&_=${DateTime.now().millisecondsSinceEpoch}';
-    _log(LogLevel.trace, '(LongPolling transport) polling: $pollUrl.');
-    final response = await _client.get(pollUrl, headers: headers);
+    _log?.call(LogLevel.trace, '(LongPolling transport) polling: $pollUrl.');
+    final response = await _client!.get(Uri.parse(pollUrl), headers: headers);
     if (response.statusCode != 200) {
-      _log(LogLevel.error,
+      _log?.call(LogLevel.error,
           '(LongPolling transport) Unexpected response code: ${response.statusCode}.');
 
       // Mark running as false so that the poll immediately ends and runs the close logic
@@ -71,15 +71,15 @@ class LongPollingTransport implements Transport {
     _receiving = _poll(_url, headers);
   }
 
-  Future<String> _getAccessToken() async {
+  Future<String?> _getAccessToken() async {
     if (_accessTokenFactory != null) {
-      return await _accessTokenFactory();
+      return await _accessTokenFactory!();
     }
 
     return null;
   }
 
-  Future<void> _poll(String url, Map<String, String> headers) async {
+  Future<void> _poll(String? url, Map<String, String> headers) async {
     try {
       while (_running) {
         // We have to get the access token on each poll, in case it changes
@@ -89,22 +89,24 @@ class LongPollingTransport implements Transport {
         }
 
         final pollUrl = '$url&_=${DateTime.now().millisecondsSinceEpoch}';
-        _log(LogLevel.trace, '(LongPolling transport) polling: $pollUrl.');
-        final response = await _client.get(pollUrl, headers: headers).timeout(
+        _log?.call(
+            LogLevel.trace, '(LongPolling transport) polling: $pollUrl.');
+        final response =
+            await _client!.get(Uri.parse(pollUrl), headers: headers).timeout(
           const Duration(milliseconds: 100000),
           onTimeout: () {
-            _log(LogLevel.warning, 'Timeout from HTTP request.');
+            _log?.call(LogLevel.warning, 'Timeout from HTTP request.');
             throw TimeoutException('A timeout occurred.');
           },
         );
 
         if (response.statusCode == 204) {
-          _log(LogLevel.information,
+          _log?.call(LogLevel.information,
               '(LongPolling transport) Poll terminated by server.');
 
           _running = false;
         } else if (response.statusCode != 200) {
-          _log(LogLevel.error,
+          _log?.call(LogLevel.error,
               '(LongPolling transport) Unexpected response code: ${response.statusCode}.');
 
           // Unexpected status code
@@ -113,15 +115,15 @@ class LongPollingTransport implements Transport {
         } else {
           // Process the response
           if (response.body.isNotEmpty) {
-            _log(LogLevel.trace,
+            _log?.call(LogLevel.trace,
                 '(LongPolling transport) data received. ${getDataDetail(response.body, _logMessageContent)}.');
 
             if (onreceive != null) {
-              onreceive(response.body);
+              onreceive!(response.body);
             }
           } else {
             // This is another way timeout manifest.
-            _log(LogLevel.trace,
+            _log?.call(LogLevel.trace,
                 '(LongPolling transport) Poll timed out, reissuing.');
           }
         }
@@ -129,12 +131,12 @@ class LongPollingTransport implements Transport {
     } catch (e) {
       if (!_running) {
         // Log but disregard errors that occur after stopping
-        _log(LogLevel.trace,
-            '(LongPolling transport) Poll errored after shutdown: ${e.message}');
+        _log?.call(LogLevel.trace,
+            '(LongPolling transport) Poll errored after shutdown: $e');
       } else {
         if (e is TimeoutException) {
           // Ignore timeouts and reissue the poll.
-          _log(LogLevel.trace,
+          _log?.call(LogLevel.trace,
               '(LongPolling transport) Poll timed out, reissuing.');
         } else {
           // Close the connection with the error as the result.
@@ -143,7 +145,7 @@ class LongPollingTransport implements Transport {
         }
       }
     } finally {
-      _log(LogLevel.trace, '(LongPolling transport) Polling complete.');
+      _log?.call(LogLevel.trace, '(LongPolling transport) Polling complete.');
 
       // We will reach here with pollAborted==false when the server returned a response causing the transport to stop.
       // If pollAborted==true then client initiated the stop and the stop method will raise the close event after DELETE is sent.
@@ -173,7 +175,7 @@ class LongPollingTransport implements Transport {
 
   @override
   Future<void> stop() async {
-    _log(LogLevel.trace, '(LongPolling transport) Stopping polling.');
+    _log?.call(LogLevel.trace, '(LongPolling transport) Stopping polling.');
 
     // Tell receiving loop to stop, abort any current request, and then wait for it to finish
     _running = false;
@@ -183,7 +185,7 @@ class LongPollingTransport implements Transport {
       await _receiving;
 
       // Send DELETE to clean up long polling on the server
-      _log(LogLevel.trace,
+      _log?.call(LogLevel.trace,
           '(LongPolling transport) sending DELETE request to $_url.');
 
       final headers = <String, String>{};
@@ -195,11 +197,12 @@ class LongPollingTransport implements Transport {
         headers['Authorization'] = 'Bearer $token';
       }
 
-      await _client.delete(_url, headers: headers);
+      await _client!.delete(Uri.parse(_url!), headers: headers);
 
-      _log(LogLevel.trace, '(LongPolling transport) DELETE request sent.');
+      _log?.call(
+          LogLevel.trace, '(LongPolling transport) DELETE request sent.');
     } finally {
-      _log(LogLevel.trace, '(LongPolling transport) Stop finished.');
+      _log?.call(LogLevel.trace, '(LongPolling transport) Stop finished.');
 
       // Raise close event here instead of in polling
       // It needs to happen after the DELETE request is sent
@@ -213,8 +216,8 @@ class LongPollingTransport implements Transport {
       if (_closeError != null) {
         logMessage += ' Error: ' + _closeError.toString();
       }
-      _log(LogLevel.trace, logMessage);
-      onclose(_closeError);
+      _log?.call(LogLevel.trace, logMessage);
+      onclose!(_closeError);
     }
   }
 }
