@@ -1,4 +1,4 @@
-class TextMessageFormat {
+mixin TextMessageFormat {
   static const RecordSeparatorCode = 0x1e;
 
   static String recordSeparator =
